+++ index.html
@@ -165,6 +165,6 @@
     </section>
   </template>
 
-  <script src="/2167717-kekstagram-simple-27/js/main.js" type="module"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
